No eta-expansion for coinductive records
Originally, we eta-expanded all records and added the eta symbol, however in agda coinductive records do not have eta-equivalence, as eta-expanding a coinductive record leads to non termination. To correct this
- When compiling the definition of a record, we check in extractRules if it has eta equality. If it is the case we proceed as usual, if not we declare the rewriting rule etaExpand _ recType x --> x, in order to eliminate the etaExpand symbol.
- On the etaExansion TCM action, we check if a term is a record without eta, in which case we do not annotate it with the symbol etaExpand.
obs. Now coinduction by copatterns is probably almost working, the only thing left is to translate correctly the copatterns