fix errors with record projection clauses
before we were always taking the clauses from funCovering, however it seems that the record projection clauses don't go through the covering check. thus, we need to check if a function is a record projection to decide if we take the clauses from funCovering or funClauses. moreover, we also need to check if a function is a real record projection or just projection-like (functions which are treated as projections due to an agda optimisation).